Add: add search form component
